Reorganise auth-non-related hooks into utils and add more test accordingly
